Partial revert: VSCR: balloon public functions with virtual_screen
This patch reverts the patch

"
VSCR: balloon public functions with virtual_screen

This patch change all public functions to use virtual_screen in balloon.h
"

Because when the balloon is painted, the vscr->screen_ptr is null and then
wmaker crash.

This patch only reverts the showBalloon function and the static funtions that
uses the virtual_screen argument used in showBallown.