Check buttons before create them in wWindowUpdateButtonImages
We should check first if the buttons are used or not, and then create
the window frame with the final buttons.

The current code creates the buttons, and then remove or add some of
them if some variables are set.

This patch check firt the variables, and then create the final buttons.