Don't hide the right button
We don't need hide this button. The button should be present. The button
was created previously and we don't need remove/hide it. If the button
is not needed, we should don't create it some lines above.

I tested and the button should be present, so, we should remove it.