wframewindow_map remove flags argument - step 1
This is the first of a patch set. The aim is remove the flags argument
in the wframewindow_map and use the flags set in the create function.

This we can use the create and map in different functions.

This first patch only adds the flags in the function wFrameWindowUpdateBorders
because this fucntion will be the only way to modify the window flags.