SF: Single functions for remove_icons_callback
This is the first of a set of patches. The idea is know what kind of
dock is used. I am using this idea to split and win the creation and
mapping of docks, clip and drawers.

This patch creates the functions drawer_remove_icons_callback and
clip_remove_icons_callback to remove the icons in the clip and the
drawers. The code come from the removeIconsCallback function.