Replace some wWorkspaceMenuUpdate with internal functions
This patch replaces some wWorkspaceMenuUpdate calls with the function
inside wWorkspaceMenuUpdate.

If we know that we are creating something, we don't need call the
update function, we need call the creating function.

The function wWorkspaceMenuUpdate updates the workspace menu, but
we know that we are creating, so, call the function that
wWorkspaceMenuUpdate uses to create the item.