wWorkspaceDelete do not realize menu
The function wWorkspaceDelete do not need refresh/realiza the window
menu, because the menu exists, but is not painted/mapped on the screen.

Is not possible to have this menu in the screen, because the menu is
removed when the user leaves it.

wWorkspaceDelete do not need run this code.