create_workspace use char pointer for wks name
This patch includes the wksname variable to set the workspace name.
The patch also checks if parr is NULL to set the value of the workspace
to NULL.

The parr variable cannot be NULL because is checked in the function
wWorkspaceRestoreState, so if is NULL is because "something is wrong".

The main idea is join both workspace create functions, and this change allow
to do it. The variable parr will be NULL if the user uses wCreateWorkspace.