create_workspace is now void
The function create_workspace should be a void function, because the
returned value is not used.

The returned value used in the function wWorkspaceNew is not used too,
because the value is not used.

I didn't change the function prototype in this patch and I add the FIXME
comment because this function should be removed and implemented as
workspace_create and workspace_map. I will do it in some patches.