SM: Observer does the map
This patch moves the map from the all functions to the observer.

All functions does the same code:

- Menu Realize,
- Menu Move if out of screen,
- Menu Paint

except switchmenu_changestate, that only paints. I override it,
becasue the code should be the same for all calls. If the change
state call doesn't need move the menu, it won't move it.