Removed function set_dockmenu_common_code
The function set_dockmenu_common_code is removed and the code is moved
to the functions:

set_dockmenu_dock_code
set_dockmenu_clip_code
set_dockmenu_drawer_code

These functions do not need the argument index, because the variable index
is moved inside the function. We need change the call *index to index, because
we are using the value, no the value in the pointer.

I moved the Wapp code to the top and the function mapping to the bottom in
the three functions. All runs ok.