New function icon_for_wwindow_miniwindow_map
The new function icon_for_wwindow_miniwindow_map sets the miniwindow
(docked app) inside the WAppIcon.

There was a bug here (since commit 1867b25513b51c443edff791bf14156ebf6eb97c):
  Date:   Sat Aug 23 11:32:01 2014 +0200
  IA: New function wAppIcon_map

Because the Application is created, the WAppIcon is created, the icon
is created, the icon->owner in the function icon_for_wwindow_map is using the
icon as owner and not the application.

This patch solves the problem, creating a new function, like
icon_for_wwindow_map but becase we know that is a miniwindow, we can set the
application as icon.