C2: Clip is not saved in screen wScreenSaveState
The clip is not saved at wScreenSaveState, because the clip belongs to the
workspace now. When the workspace is saved, the clip (inside the workspace)
is saved too.

This patch removes the "common clip" for all workspaces and allow have clips
with different Monitors in different positions.

The patch also removes the not used functions and old code.