SR4: Do not map the clip in the creation function
This patch do not map the clip in the creation function. We can
remove the function set_clip_in_workspace_map2 because the clip
will be painted later, in the map function.

The patch includes a bug. Some patches before, I detected the docked
icons to the Clip are moving from one workspace to the previous
workspace when wmaker restarts. For example, the icon is attached to
the clip at workspace 3, when restarts, the icon is the workspace 2,...
But this patch includes that the icon when is in workspace 1 and awmaker
is restarted, the icon is dropped. Finally, the icon is painted always
in the workspace 1. When the workspace is changed, the icon are refreshed
and all is fine.

I will solve this bugs later, but this change is very important and is
better do not include more changes here.