Icons include virtual screen 1/3
This is a set of patches to include the virtual screen in the icons,
dockapps, appicons. All pathes must be applied together, but I split
them to avoid a big patch.

This is the first patch, it includes changes in the function create_appicon
and other functions must be changed too.