Icons include virtual screen 2/3
This is a set of patches to include the virtual screen in the icons,
dockapps, appicons. All pathes must be applied together, but I split
them to avoid a big patch.

This is the second patch, it includes changes in the function dock_icon_create
and in the function clip_icon_create. Other functions must be changed too.