Icons include virtual screen 3/3
This is a set of patches to include the virtual screen in the icons,
dockapps, appicons. All pathes must be applied together, but I split
them to avoid a big patch.

This is the third patch, it includes changes in the function icon_create_core.
Other functions must be changed too.