Look for plmenu in build directory when creating WMRootMenu.
Based on the patch from Doug Torrance <dtorrance@piedmont.edu>. Rewritten for
awmaker.

Previously, plmenu already existed in the source directory, and that's where
make looked for it.  Now, however, it is generated to include the proper
path to WPrefs and therefore is in the build directory.

See
https://www.gnu.org/software/make/manual/html_node/Makefile-Basics.html
for more about this distinction.