Implementation for moving mouse pointer within the maximized window.
Based on the patch from Roman Dobosz <gryf73@gmail.com>. Rewritten for
awmaker.

Mouse pointer can be now moved together with window if keyboard was used
for arrange maximized windows on screen. This feature can be turned on
in WPrefs.app in Expert tab by selecting "Move mouse pointer with half
MAXIMIZED windows.", or setting "PointerWithHalfMaxWindows" to "Yes" on
~/GNUstep/Defaults/WindowMaker file.