wmaker: allow alt+tabbed windows over fullscreen
Based on the patch from Bjørn Mork <bjorn@mork.no>. Rewritten for
awmaker.

Fullscreen windows should only be on top when they are in focus. Change
the stacking level temporarily back to WMNormalLevel if the fullscreen
window loses focus due to an alt+tab operation.

Change the stacking level back to WMFullscreenLevel if the fullscreen
window receives the focus again.